Create demo WorkingExample
