+++ src/App.js
@@ -1,31 +1,12 @@
 import React from 'react';
 import './App.css';
-import { useDispatch, useSelector } from 'react-redux';
-import { logIn, logOut } from './reducers/slices/login';
+import WorkingExample from './reducers/WorkingExample';
 
 function App() {
   
-  // Get state from store with useSelector and useDispatch.
-  const state = useSelector(state => state.auth); // State with the name 'auth'
-
-  const dispatch = useDispatch();
-
-  // Dispatch logIn and logOut depending on state.isAuthenticated.
-
-  function handleClick() {
-    dispatch(!state.authData.isAuthenticated?logIn(state.authData.isAuthenticated): logOut(!state.authData.isAuthenticated));
-    console.log(state.authData.isAuthenticated);
-    
-  }
-
-  // Click event to dispatch logIn and logOut.
-
   return (
     <div className="App">
-      <header className="App-header">
-        <button onClick={handleClick}>LogIn or LogOut?</button>
-        {state.message}
-      </header>
+      <WorkingExample />
     </div>
   );
 }
